ui: refactored yaml editor to being generic, further stabilised feature
